+++ client/cache_rule_test.go
@@ -66,3 +66,4 @@ func TestCreateCacheRule(t *testing.T) {
 		t.Fatal(err)
 	}
 }
+