+++ prosimo/resource_app_onboarding_cloud_svc.go
@@ -323,6 +323,12 @@ func resourceAppOnboarding_CloudSVC() *schema.Resource {
 				Default:     true,
 				Optional:    true,
 			},
+			"force_offboard": {
+				Type:        schema.TypeBool,
+				Description: "Force app offboarding incase of normal offboarding failure.",
+				Default:     true,
+				Optional:    true,
+			},
 		},
 		Timeouts: &schema.ResourceTimeout{
 			Create: schema.DefaultTimeout(60 * time.Minute),