+++ prosimo/resource_app_onboarding_dns.go
@@ -244,6 +244,12 @@ func resourceAppOnboarding_DNS() *schema.Resource {
 				Default:     true,
 				Optional:    true,
 			},
+			"force_offboard": {
+				Type:        schema.TypeBool,
+				Description: "Force app offboarding incase of normal offboarding failure.",
+				Default:     true,
+				Optional:    true,
+			},
 		},
 		Timeouts: &schema.ResourceTimeout{
 			Create: schema.DefaultTimeout(60 * time.Minute),